Screen off battery alerts [1/3]
This alerts when the screen off and charging if the charge level
becomes good.

Don't alert if screen is on, or it's been off since less than one
minute, or if battery isn't charging.

Show a notification (good e.g. for Ambient display if you don't hear the sound).
Play a sound if ringer mode is normal, otherwise vibrate if ringer mode is vibrate,
otherwise nothing.

Notification and sound/vibrate code are separated so the user can even
disable the notification (long tap on it) but stil keep sound/vibrate alert.

Thanks to @xdnax for the nice idea and suggestions

Change-Id: I1c300ba94ac01ab42dd2585144830727b4da092d

Signed-off-by: Robby Duran <storm119@gmail.com>